# openapi/dazat-api.yaml
openapi: 3.1.0
info:
  title: DAZAT Mobility API
  version: 0.0.6
  description: Engineering Phase 0.6 API. Server/domain owners remain authoritative for identity, Booking, Dispatch, Journey telemetry, governed changes, Safety and completion evidence.
servers:
  - url: http://localhost:3001
paths:
...
  /v1/safety/signals/sos:
    post:
      summary: Persist an SOS as a restricted INCIDENT signal before any external delivery attempt.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JourneySafetySignal' }
      responses:
        '201': { description: Canonical Safety event and restricted outbox record committed. }
  /v1/safety/signals/silent-assistance:
    post:
      summary: Persist Silent Assistance with a mandatory do-not-auto-call-reporter policy.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JourneySafetySignal' }
      responses:
        '201': { description: Signal committed locally first; external provider availability is not a persistence dependency. }
  /v1/safety/route-concerns:
    post:
      summary: Record a contextual route concern without creating an automatic misconduct finding.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteConcern' }
      responses:
        '201': { description: Restricted Safety event plus contextual route-deviation signal committed. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
...
  schemas:
...
    JourneySafetySignal:
      type: object
      required: [journeyId]
      properties:
        journeyId: { type: string, format: uuid }
    RouteConcern:
      allOf:
        - $ref: '#/components/schemas/JourneySafetySignal'
        - type: object
          required: [routeConcernCategory]
          properties:
            routeConcernCategory: { type: string, enum: [CHECK_ROUTE, WRONG_DESTINATION, FEEL_UNSAFE, UNEXPECTED_STOP, OTHER] }